+++ packages/cli/src/app/services/legacy/NwayoLegacyService.js
@@ -1,5 +1,5 @@
 //--------------------------------------------------------
-//-- Nwayo - Providers - Application Service Provider
+//-- Nwayo - Legacy - Nwayo Legacy Service
 //--------------------------------------------------------