+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Changed hover selection from 'mouse' to 'not-touch'
+- Changed jpeg optimization from jpegtran to MozJPEG
+- Update to Modernizr 3.9.0
+- Maintenance updates